select rpc component - rename file
